requirements.txt
Flask
Pillow
tensorflow
numpy
opencv-python-headless
mtcnn
google-generativeai